ai-network-autoconfig/requirements.txt
netmiko
pyyaml
GitPython
openai
